Create ProfilerBase
